requirements.txt
# Core dependencies
torch>=2.0.0
diffusers>=0.24.0
transformers>=4.30.0
accelerate>=0.21.0
safetensors>=0.3.2
Pillow>=9.5.0
numpy>=1.24.0
matplotlib>=3.7.0
tqdm>=4.65.0

# For IP-Adapter (optional - could also be installed directly from GitHub)
# pip install git+https://github.com/tencent-ailab/IP-Adapter.git

# For LoRA fine-tuning
peft>=0.5.0

# For text generation
sentencepiece>=0.1.99

# Utilities
scipy>=1.10.0
requests>=2.31.0
ftfy>=6.1.1
regex>=2023.6.3
